requirements.txt
streamlit>=1.28.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=0.8.11
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
alembic>=1.12.0
bcrypt>=4.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
